backend/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
onnx
roboflow